code/requirements.txt
django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
django-environ==0.12.0
django-filter==24.3
requests==2.31.0
PyYAML==6.0.2
jsonschema==4.25.1
Pillow==10.4.0
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10

# RAG (Retrieval-Augmented Generation) Dependencies
chromadb==0.4.22
sentence-transformers==2.3.1
langchain==0.1.20
langchain-community==0.0.38
langchain-chroma==0.1.2
tiktoken==0.6.0
